requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.17
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.1
dill==0.3.8
eval_type_backport==0.2.0
evaluate==0.4.3
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
huggingface-hub==0.25.1
idna==3.10
imageio==2.35.1
Jinja2==3.1.4
kaggle==1.6.17
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.10.0.84
opendatasets==0.1.22
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pynvml==11.5.3
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
sympy==1.13.3
text-unidecode==1.3
tifffile==2024.9.20
timm==0.5.4
tokenizers==0.20.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.45.1
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
xxhash==3.5.0
yarl==1.13.1
